# packages/types/python/user.py
# packages/types/python/user.py
# T010 [P] Define packages/types/python/user.py based on User entity

from pydantic import BaseModel, Field

class UserBase(BaseModel):
    username: str = Field(..., min_length=3, max_length=50)

class UserCreate(UserBase):
    password: str = Field(..., min_length=8) # This will be hashed

class UserInDB(UserBase):
    id: int
    password_hash: str

    class Config:
        from_attributes = True

class UserPublic(UserBase):
    id: int

    class Config:
        from_attributes = True